extraction/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
# Bump libraries to versions compatible with Python 3.13
pdfplumber==0.11.4
pytesseract==0.3.10
Pillow==11.0.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic-core==2.23.4
pydantic==2.9.2
opencv-python==4.10.0.84
numpy==2.1.2
python-dateutil==2.8.2
